+++ packages/spacecat-shared-utils/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [@adobe/spacecat-shared-utils-v1.123.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.123.0...@adobe/spacecat-shared-utils-v1.123.1) (2026-06-26)
+
+### Bug Fixes
+
+* **utils:** added some patterns for aem_headless detection ([#1716](https://github.com/adobe/spacecat-shared/issues/1716)) ([a646b2b](https://github.com/adobe/spacecat-shared/commit/a646b2bf8aeb0bb11e13e03c43f57f414af7caad))
+
+## [@adobe/spacecat-shared-utils-v1.123.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.122.1...@adobe/spacecat-shared-utils-v1.123.0) (2026-06-26)
+
+### Features
+
+* **rum:** scope latest-metrics RUM to locale path prefix ([#1707](https://github.com/adobe/spacecat-shared/issues/1707)) ([77cdb33](https://github.com/adobe/spacecat-shared/commit/77cdb33d7f55a9bff795b768621ec70830b2cba3))
+
 ## [@adobe/spacecat-shared-utils-v1.122.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.122.0...@adobe/spacecat-shared-utils-v1.122.1) (2026-06-26)
 
 ### Bug Fixes